BabelNet Synset

Retrieve BabelNet synset information.

Usage

bablenet_synset(ids, outDir = NULL, overwrite = FALSE,
  cache = paste0(dirname(tempdir()), "/babelnet_pages/"),
  cores = parallel::detectCores() - 2)

Arguments

ids

A vector of BabelNet IDs (e.g., "bn:00103811a").

outDir

Path of a directory in which to save parsed page JSON files.

overwrite

Logical; if TRUE, will overwrite existing parsed pages when retrieved.

cache

Path to a directory in which to save original page HTML files.

cores

Number of CPU cores to use during processing

Value

A list with entries for each id in ids.

Examples

(bablenet_synset("bn:00103811a"))
#> $`bn:00103811a`
#> $`bn:00103811a`$id
#> [1] "bn:00103811a"
#> 
#> $`bn:00103811a`$lemmas
#> [1] "gross"   "porcine"
#> 
#> $`bn:00103811a`$definitions
#>                             source         content
#> 1 WordNet 3.0|Open English WordNet Repellently fat
#> 
#> $`bn:00103811a`$examples
#>                             source                content
#> 1 WordNet 3.0|Open English WordNet A bald porcine old man
#> 
#> $`bn:00103811a`$relations
#> $`bn:00103811a`$relations$SIMILAR
#>             id term
#> 1 bn:00102820a  fat
#> 
#> 
#> $`bn:00103811a`$sources
#> $`bn:00103811a`$sources$`WordNet 3.0 &amp; Open English WordNet`
#>      name         type                                                  url
#> 1   gross HIGH_QUALITY   http://wordnetweb.princeton.edu/perl/webwn?s=gross
#> 2 porcine HIGH_QUALITY http://wordnetweb.princeton.edu/perl/webwn?s=porcine
#> 
#> 
#>