Get Baby Names Data

Downloads and processes historical baby names data from the U.S. Social Security Administration.

Usage

get_baby_names(dir = tempdir(), source = "national")

Arguments

dir

Directory in which to save data.

source

Which dataset to download; between national (default), state, and territory.

Value

A list with entries for data (a data.frame with the original data), and summary (a matrix with a count of each name and probabilities per sex).

Examples

if (FALSE) { # \dontrun{
# saves to a temporary directory
names <- get_baby_names()
} # }