Guess sex based on a web search

Searches for a name (or whatever), and looks for gendered pronouns or other basic cues to sex.

Usage

guess_sex(names, retry = 50,
  search_source = "https://search.yahoo.com/search?q=",
  cores = detectCores() - 2)

Arguments

names

A character vector of search terms (such as a name).

retry

Logical; if TRUE, will retry after a failed attempt.

search_source

URL of the search engine; each entry in names is simply appended.

cores

Number of CPU cores to split names across.

Value

a data.frame with entries for name (the original search term), guess, confidence, and female and male with count of matches.

Examples

if (FALSE) { # \dontrun{
guess_sex(c("Jane", "John"))
} # }