lusilab: Language Use and Social Interaction Lab

Includes functions for power analysis, permuting or bootstrapping mediations, automated factor analysis, and pulling text from reddit, imsdb, and rottentomatoes, plus some other things to play with.

See also